// src/models/Historical.ts
// Generated by https://quicktype.io

export interface ExchangeRateHistoricalDateResponse {
    motd?:       MOTD;
    success:    boolean;
    historical: boolean;
    base:       string;
    date:       Date;
    rates:      { [key: string]: number };
}

export interface MOTD {
    msg: string;
    url: string;
}